Implement the initial interface CloudProvider
This commit implements the initial interface for cloud providers.

The idea is to have the `CloudProvider` as a common interface for
uploading and publishing a VM Image during the `push` workflow.

With this it's possible to the task implementation to simply "retrieve"
the correct provider with `CloudProvider.get_provider()` and then
calling the `upload()` and `publish()` method for each push item.

Refers to EXDSP-1809