Fix CloudProvider interface and Azure SAS
This commit changes the CloudProvider interface to always return the
Push Item as we may need to change it in the post actions.

It also fixes the AzureProvider to change the VHDPushItem on
`_post_upload` step to include the SAS URI in the push item, otherwise
it will be impossible to publish it as the value would be None.

Finally, update the tests to reflect the changes.