Don't pass missing disk_version to CloudPub
In order to associate an image with an offer/plan on Azure it's
necessary to provide the DiskVersion.version number which is a string
expecting the format `{int}.{int}.{int}`. Without this value it's
impossible to proceed with publishing.

In order to solve this issue, as today, Feb 17, 2023, the CloudPub
library defaults to `1.0.0` when the disk_version is not set, which is
not the desired solution.

This commit implements a new method on AzureProvider named
`_generate_disk_version` which will automatically generate the required
string using the definitions from our design requirements.