Prevent multiple publishing for the same offer
On AWS all product listings has a unique destination as they come with
an UUID.

For Azure, however, it's possible to have multiple destinations for a
single offer, as shown below:

- offer/plan1
- offer/plan2

In this case the same offer has 2 plans. It does make sense to update
this offer twice as a draft to associate the images to each plan,
however it does not make sense to publish it to live twice.

This commit prevents the publishing to invoke the `send to live` state
multiple times for the same offer when pre-push == False.