Fix bug on marketplaces destinations mapping
The expected behavior of our pubtool is to publish the Push Items to all
destinations for the given marketplaces.

Before this fix the pubtool was pushing ALL destinations for every
marketplace, which is not the desired behavior. So, considering the
following example:

  aws-na:
    - offer1
  aws-emea:
   - offer2

It was pushing to `offer1` and `offer2` to both `aws-na` and `aws-emea`
instead of just `offer1` to `aws-na` and just `offer2` to `aws-emea`.

This commit fixes this issue so that the pubtool will only publish to
the destinations corresponding to the actual marketplace being processed