Trim minutes on Azure disk version
During the tests for demo I found out that Azure can't handle a disk
version with the last integer having a length of 12 digits but it
accepts up to 10 digits.

Before this commit the disk version was being generated in the format
below:

`{major}.{minor}.{YYYYmmddHHMM}`

Being `{YYYYmmddHHMM}` an integer with 12 digits representing the
current year, month, day, hour and minute.

This commit changes it to:

`{major}.{minor}.{YYYYmmddHH}`

in order to respect the 10 digit length imposed by Azure.