Allow uploading image to custom container from map
This commit allows the `pubtools-marketplacesvm` to upload the VM images
to a custom container from a StArMap mapping whenever the `meta`
attribute `upload_container` is set.

It keeps the default `pubupload` from `UPLOAD_CONTAINER_NAME` whenever
the attribute is not set from the MappedVMIPushItem (StArMap mapping).

With this it's possible to properly upload the artifacts to AWS using
different S3 buckets for each marketplace account (EMEA/NA).

Refers to EXDSP-2061