Only upload images during nochannel phase
This commit changes the `_push_to_cloud` behavior so it will only upload
the VM image if the request comes with `pre_push == True` instead of
uploading and associating the images.

This change is required since we can no longer leave the offer in the
"draft" state as any offer in this original state will be considered
"not good" to proceed, it is, it will represent that someone made manual
changes in the offer through the Web UI.

Whenever the `pre_push == False` it will run the `_publish` initially
with "keep_draft == True` to associate the images to possible multiple
plans and finally call the `keep_draft == False` to submit changes live.

Refers to EXDSP-2080