Prevent AzureProvider to change "draft" offers
This commit prevents the `AzureProvider` to modify offers which initial
state is "draft" whenever `nochannel is True`.

During a pre-push operation (nochannel) we don't want to modify an offer
which is in `draft` state as it may represent some manual changes done
in the Web UI directly.

Since the `pubtools-marketplacesvm` workflow for Azure on `publishing`
requires the draft state during the image association step we use a
`borg` object to keep track the destinations we already validated.

Refers to EXDSP-2080