Allow using custom tags when uploading
This commit changes the `CloudProvider` interface by adding an optional
parameter on `upload` and `_upload` named `custom_tags` which consists
of a dictionary with key/values to be added alongside the default tags
during the upload.

It also updates `AWSProvider` and `AzureProvider` to support this
feature as well as the `push.command` to send the custom tags from
`MappedVMIPushItem` to the proper `CloudProvider`.

Refers to EXDSP-2299