Allow publishing to preview on pre_push
This commit changes the `pubtools-marketplaces` workflow to allow
publishing a VM image to the Marketplace as `preview` during the
`--pre-push` operation only and if only the StArMap flag
`stage_preview` is set to True.

The reason for this change is because QE validation for Azure needs
the product in the marketplace when these are not RHEL. It will be
under responsibility of release engineer to proper maintain the
StArMap Policy Mapping correct.

Note that this change doesn't apply to AWS as it's not possible to
publish a VM image in "preview" for this marketplace, so the new
workflow is just "ignored" by AWSProvider

In order to implement this new workflow the following changes were
made:

- Bump `cloudpub` and `starmap-client` to use the new StArMap
  Destination's attribute `stage_preview` as well as be possible
  to publish to "preview" in Azure.

- Change the CloudProvider's `publish` and `_publish` interfaces
  to have a new optional attribute named `preview_only` which will
  inform cloudpub to perform the publish to `preview`

- Change the `Push` workflow to send the `stage_preview` attribute
  from StArMap to the `_publish` method.

Depends on:
  - cloudpub!56
  - starmap-client!18

Refers to EXDSP-2463