BUGFIX: Use the correct mapping for publishing
This commit fixes a bug on publishing which was using a general mapping
instead of the one for the current marketplace. It also enforces the
MappedPushItem to use `get_push_item_for_marketplace` by removing the
read access to the property `push_item`, which should be only handle
internally.

With this the push items will be able to be published with different
offers/plans depending on the marketplace mappings.