EXDSP-2539: Fix regression with AMI ID and SAS URI
This commit fixes a bug which was preventing the AMIs IDs and SAS URIs
to be properly mapped.

Due to changes in !93 the information from `_post_upload` actions were
being overriden by the last loop stage from upload, since the original
`MappedVMIPushItem` was just holding a single instance of the push item.

This fix includes an internal dictionary for `MappedVMIPushItem` named
`_mapped_push_item` which holds information of a push item for a
specific marketplace location.

It also remove no longer needed properties from `MappedVMIPushItem`.

Finally, it includes 2 new unit-tests to prevent this issue to happen
again.

Refers to EXDSP-2539