Allow reading AmiRelease / VMIRelease from StArMap
This commit changes the `MappedVMIPushItem._map_push_item` to
create/merge the `push_item.release` data when set by StArMap using
`meta.release`.

The behavior of mapped `push_item.release` is now:

- If the `meta.release` is present in StArMap:
  - If the `push_item.release` doesn't exist it will receive all data
    from StArMap.
  - If the `push_item.release` already exists it will be updated with
    StArMap's data
- If the `meta.release` is not present in StArMap:
  - No changes will be made to `push_item.release`

With this we ensure the `pubtools-marketplacesvm` is able to read the
`release` data from StArMap and map it into the push item.

Refers to EXDSP-2522