Prepare the CloudProvider for Community workflow
This commit originally intended to adjust the `AWSProvider` to support
both the `stratosphere` and `community AMIs` workflow.

The changes are:

- Expose the `name_from_push_item` out of `AWSProvider` class to be a
  module's function so it can be used outside `AWSProvider` as well
- Allow passing the `groups` by parameter
- Including the `billing_codes` whenever the information is available

However, in order to pass the optional parameter `groups` to
`AWSProvider` we need to change the interface as well. In order to avoid
adding some specific AWS stuff to the interface and increase the
parameters to just some specific providers I decided to include the
`**kwargs` in all `upload` interfaces (`upload`, `_upload` and
`_post_upload`) so if any additional parameter is required by some
specific cloud provider it won't require changing the interface again.

This commit also updates the unit-tests to support optional keyword
arguments in the `upload` interface.

Refers to EXDSP-2528