Allow using custom S3 buckets
This commit changes the AWSProvider to allow passing a custom S3 bucket
into the `upload` parameter to override the default `self.s3_bucket`.

It also changes the `CommunityVMPush` which now has a new CMD argument
called `--container-prefix` which is used to inform the container prefix
as the final one will be composed by "{container_prefix}_{region}".

This is required in order to upload the AMIs to multiple regions as
without the proper S3 per regions we receive the following error:

```
ERROR: Bucket <MyBucketName> is not in the <RegionName> Region,
 it's in <RegionName>
```

See https://docs.aws.amazon.com/vm-import/latest/userguide/vmimport-troubleshooting.html
for details.

This commit is based on the approach used in the pubtools-ami.