Ensure the proper AMI names
This commit changes the `name_from_push_item(aws_push_item)` function
to use the existing convention from Community AMIs when this workflow is
being used.

As the main difference between stratosphere names and community names is
that the latter adds two parts: `billing_codes.name` and `volume` the
required change is possible to be done by verifying whether the
push_item is meant for community workflow or not.