Prepare pubtools for the unified workflow
This commit changes the `run` method interface for `MarketplacesVMTask`
and its subclasses as well as adjust this method to be compatible with
the unified workflow.

The goal here is to be able to add in future a third endpoint which will
be responsible to trigger the `stratosphere` and `community` workflow in
the same push while controlling the data for collection and allowing or
not one or more workflows to skip items (if no mappings are defined).

The new `run` interface has the following optional arguments:

- `collect_results` which defaults to `True`. When set to `False` it will
  skip the result collection on `pushcollector`, so we can execute this
  call once for both workflows in the execution of `CombinedVMPush`.

- `allow_empty_targets` which defaults to `False`. When set to `True` it
  will not fail when StArMap doesn’t have mappings for the incoming
  push items.

The return value `RUN_RESULT` is a named tuple with the following data:

- `success` A boolean to indicate whether the push was successful or not.
  Note that when allow_empty_targets is set to True ** it will return `True`
  even if no items were processed due to lack of mappings.

- `skipped` An auxiliary boolean to indicate whether one or more push items
  didn’t have any mappings on it. The idea behind this boolean is to evaluate
  whether both `MarketplacesVMPush` and `CommunityVMPush` returned `False` so
  it will raise an error (indicating no mappings for both workflows).

- `collected_result` A dictionary containing the data for `pushcollector`,
  so the upcoming class `CombinedVMPush` can combine them to send once
  for the collection.

Refers to SPSTRAT-244