Use a common `sharing_accounts` format for AWS
This commit changes the `community` workflow to implement the support
for AWS `sharing_accounts` meta attribute from StArMap in the same way
as it's being used in the `marketplace` workflow.

As a retrocompatibile measure the patch still allows the use of the
previous format exclusive to the `community` workflow which can be set
by the meta attribute `accounts`.

Refers to SPSTRAT-303