Implement RepoQueryLoad action
This commit implements a new ArgParse action named `RepoQueryLoad` which
is responsible to load a JSON string from argument and convert it to a
list of StArMap's `QueryResponse` objects.

As a consequence of parsing the data as `QueryResponse` we also get the
validation from the `starmap_client` library.

It will be used to replace the way we load data from `--repo`.

Refers to SPSTRAT-323