Update repo argument to load QueryResponse
This commit changes the stratosphere workflow when overriding the
StArMap queries, as it will now receive a complete QueryResponse from
the `--repo` argument instead of a dictionary.

Note that even though we're now using the `RepoQueryLoad` action, which
creates a list of QueryResponse objects, it's still not possible to have
multiple mappings definition as we will implement the management of such
list in the `starmap_client` library.

What we have now is just a drop in replacement for the existing `--repo`
feature which behaves the same way but receives a different format from
its input.

Refers to SPSTRAT-323