Accept a list of mappings in --repo
This commit changes the `--repo` argument to be part of StArMap service,
allowing the tooling to receive a list of mappings by command line which
will be used in replacement of the StArMap server.

With this, whenever a given mapping is found it will not query the
server.

Since this change moves the `--repo` to the `StarmapService` it now
applies the same effect to all workflows, which was not supported
before.

Refers to SPSTRAT-324 and SPSTRAT-325