Allow using Starmap Client Offline Mode
This commit introduces a new flag on `starmap_service` named
`--offline`, which in conjunction with `--repo` allows the pubtools to
not request any mapping in the StArMap service.

The idea is to be able to inject mappings using `--repo` and do not send
any requests to the server for not defined mappings, allowing the
execution of the tooling in a network isolated environment.

Refers to SPSTRAT-326