Enhance the empty workflow detection
Some products, like RHCOS, generates multiple push items to a single
build. In RHCOS case, it generate both `VHDPushItem` and `AmiPushItem`
objects.

However, there are some scenarios which we want to publish some content
just to a given marketplace OR a given workflow. In the previous
versions, this was not possible, as detailed below:

Suppose a build for RHCOS which generates both push items for Azure and
AWS and we just want to publish it on AWS community AMIS.. The following
conditions would make the push fail:

- marketplace_ami: no mappings - community_ami: pushed
- marketplace_azure: no mappings - community: no mappings <<<< FAIL

The second push would trigger a failure since the `VHDPushItems` for
RHCOS didn't get pushed to any workflow (community or marketplace).

As a way to properly fix it, this commit introduces a borg named
`BuildIdBorg` which will be responsible to track the build IDs in the
incoming of a push item and after its processing.

On the end of the combined workflow we check whether both workflows did
skip AND whether there's a build id which wasn't processed before
raising errors.

With this, the RHCOS scenario no longer gives errors as the build ID is
the same for all its push items.

Refers to SPSTRAT-309