Refactor the community threads workflow
This commit fully refactors the `community_push` threading workflow by
creating a single threadpool for consuming the required data to call the
`_upload` method and doing it in parallel.

It also creates a generator method named `_data_to_upload` which
provides the required data to upload the push items.

With that, we have a streamlined operation which spawns multiple threads
to process in parallel each push item and its region in an optimized
way.