Community: Fix sharing accounts bug
This commit fixes a bug which was causing all push items from community
workflow to receive the very same sharing accounts (from the first load)
instead of receiving the proper onces per account provider.

The fix consists on changing the `EnrichedPushItem` structure to hold
now a tuple containing the push item and its related sharing accounts,
which are now being retrieved from the specific `Destination` object of
that push item.

Refers to SPSTRAT-356