Add a proper method to get PI from a Destination
This commit changes the `MappedVMIPushItem` by adding a new method named
`get_push_item_for_destination` which returns a mapped push item using
the data from the given destination, instead of a list of destinations.

This change is required to comply with pushitems which are expected to
have different property values depending on its destination, it is, the
`meta` data from each destination may have a different value.

As a consequence of adding this method we split the internal method
`_map_push_item` into two:

1. `_map_push_item`: Will just check if the internal pushitem was mapped
   and if not call the `_update_push_item_properties`
2. `_update_push_item_properties`: Will update the properties of a given
   push item.

It also updates the community push to use the new method
`get_push_item_for_destination` as their pushitems are expected to have
different values from each Destination.