DeleteCommand: Fix delete_push_images interface
This commit changes the `delete_push_images` interface to always return
a tuple with the push item and the deletion result.

It changes the `AWSProvider._delete` method to return the deletion
result as well as fixes the return object of `_delete_push_images` which
expects the push_item and not the type.

Refers to SPSTRAT-454