Allow deletion to continue if RHSM update fails
This commit enhances the `DeleteCommand` by allowing the deletion to
continue whenever there's an error on RHSM update to mark the image as
invisible.

It also improves the logging to inform the RHSM operation to make the
AMI invisible is about to start and shows the error when it occurs.

Refers to SPSTRAT-454