Add unit-test for multiple arches destinations
This commit adds a unit-test to reproduce a situation when some push
items got different destinations depending on its architecture.

The current version on `main` has a bug which is merging the
destinations of one architecture into another ones.

The bug is fixed by this PR's changes and it's a good practice to add
the unit test so future changes may be tested alongside it.

Refers to SPSTRAT-533

Signed-off-by: Jonathan Gangi <jgangi@redhat.com>