Azure: Pass the partial SAS arg to pre-publish
This commit is a bugfix related to !107. Since the tooling will first
associate the VHD image with the plan and then publish, it must pass the
argument `check_base_sas_only` to the `pre_publish` mehotd or it won't
perform the partial SAS check and fail on arguments comparison.

Refers to SPSTRAT-547

Signed-off-by: Jonathan Gangi <jgangi@redhat.com>