bugfix: Azure delete: return response
This commit fixes an issue with the `delete` command for Azure which was
failing with the following error:

```
     pi, res = self._delete_push_images(push_item, **kwargs)
     ^^^^^^^
TypeError: cannot unpack non-iterable VHDPushItem object
```

The issue was caused because the interface expects a tuple to be
returned from `_delete_push_images` and the aforementioned method was
only returning the `push_item` for AzureProvider.

It also adds a full end-to-end test for Azure deletion.

Refers to SPSTRAT-610

Signed-off-by: Jonathan Gangi <jgangi@redhat.com>